Added Unit Test target to Xcodegen project.yml
- Renamed xcode project to SwiftDP-App
- Moved Carrots Demo to the Unit Test
