+++ .gitignore
@@ -4,6 +4,6 @@
 /bazel-testlogs
 
 # generated by XcodeGen
-/xcode/SwiftDP-Test-Framework.xcodeproj
+/xcode/SwiftDP-App.xcodeproj
 /xcode/dummy
 /xcode/build
