+++ start.sh
@@ -15,5 +15,5 @@ echo "==========================================="
 echo ""
 echo "Setup Complete 🚀"
 echo "Run:"
-echo "open ./xcode/SwiftDP-Test-Framework.xcodeproj"
+echo "open ./xcode/SwiftDP-App.xcodeproj"
 echo ""
